Changed TileMapMesh to use the new TileNode
Now only faces not adjacent to other tiles will be drawn.
The new system seemed to cut down load time, and at 100x75 tiles
framerate didn't seem to be an issue.

I was NOT able to display a 100x100 tile plain. An can't access property
'length' error gets thrown in the method that calculates the vertex
normals. As there seems to be no problem for arbitrary sizes at 100x75
and under, I'm going to chalk this up to some kind of weird out of mem.
error.