created is_adjacent_flush, fixed SN coord system
the is_adjacent_flush method returns ture if an adjacent face is flush
(occupies the same coordinates/plane). Will be used to determine whether
or not to add verts of a face that will not be visible.

Changed erroneous interpretation of tile coordinate system where north is
negative. To keep the notion of North is up on the Y axis, we need to make
South be in the negative Y direction and North in the positive.